requirements.txt
requests==2.27.1
Flask==2.0.3
Werkzeug==2.0.3
redis==3.5.3
tqdm==4.46.0
flask-cors==3.0.10
pyOpenSSL==22.1.0
psycopg2-binary
loguru==0.6.0
